requirements.txt
# Python依存パッケージ

# X (Twitter) API
requests>=2.31.0

# RSS解析
feedparser>=6.0.10

# YAML設定ファイル読み込み
PyYAML>=6.0.1

# 日付処理
python-dateutil>=2.8.2

# GitHub API（オプション: requestsでも可）
PyGithub>=2.1.1

# Claude API（X投稿素案生成用）
anthropic>=0.34.0

# X API OAuth認証（投稿用）
requests-oauthlib>=1.3.1
